feat: refactor getSubPeriodData
